+++ collector-amazon-ivs/src/main/java/com/bitmovin/analytics/amazon/ivs/api/IAmazonIvsPlayerCollector.kt
@@ -13,13 +13,12 @@ import com.bitmovin.analytics.utils.Util
 /**
  * Analytics Collector for the Amazon IVS Player
  */
-interface IAmazonIvsPlayerCollector : AnalyticsCollector<Player> {
-
+public interface IAmazonIvsPlayerCollector : AnalyticsCollector<Player> {
     /**
      * The [sourceMetadata] which is used to enrich the analytics data with source specific
      * metadata.
      */
-    var sourceMetadata: SourceMetadata
+    public var sourceMetadata: SourceMetadata
 
     /**
      * CustomData accessor to the current configured [sourceMetadata]
@@ -32,16 +31,16 @@ interface IAmazonIvsPlayerCollector : AnalyticsCollector<Player> {
      * More information can be found here:
      * https://developer.bitmovin.com/playback/docs/how-can-values-of-customdata-and-other-metadata-fields-be-changed
      */
-    var customData: CustomData
+    public var customData: CustomData
 
-    companion object Factory {
+    public companion object Factory {
         /**
          * The version of the analytics collector
          *
          * For example "3.0.0"
          */
         @JvmStatic
-        val sdkVersion = Util.analyticsVersion
+        public val sdkVersion: String = Util.analyticsVersion
 
         /**
          * Creates a collector instance configured via the provided [analyticsConfig], and the
@@ -49,7 +48,11 @@ interface IAmazonIvsPlayerCollector : AnalyticsCollector<Player> {
          */
         @JvmStatic
         @JvmOverloads
-        fun create(context: Context, analyticsConfig: AnalyticsConfig, defaultMetadata: DefaultMetadata = DefaultMetadata()): IAmazonIvsPlayerCollector {
+        public fun create(
+            context: Context,
+            analyticsConfig: AnalyticsConfig,
+            defaultMetadata: DefaultMetadata = DefaultMetadata(),
+        ): IAmazonIvsPlayerCollector {
             val collector = AmazonIvsPlayerCollector(analyticsConfig, context)
             collector.defaultMetadata = defaultMetadata
             return collector