+++ collector/src/main/java/com/bitmovin/analytics/Observable.kt
@@ -1,6 +1,10 @@
 package com.bitmovin.analytics
 
+import com.bitmovin.analytics.internal.InternalBitmovinApi
+
+@InternalBitmovinApi
 interface Observable<TListener> {
     fun subscribe(listener: TListener)
+
     fun unsubscribe(listener: TListener)
 }