+++ collector/src/main/java/com/bitmovin/analytics/OnAnalyticsReleasingEventListener.kt
@@ -1,5 +1,8 @@
 package com.bitmovin.analytics
 
+import com.bitmovin.analytics.internal.InternalBitmovinApi
+
+@InternalBitmovinApi
 interface OnAnalyticsReleasingEventListener {
     fun onReleasing()
 }